requirements.txt
aiofile==3.9.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.7.2
async-timeout==4.0.3
asyncssh==2.18.0
attrs==24.2.0
bcrypt==4.0.1
beautifulsoup4==4.13.3
billiard==4.2.0
black==24.4.2
bs4==0.0.2
cachetools==4.2.2
caio==0.9.17
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.0
coveralls==4.0.1
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.11
django-cors-headers==4.7.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docopt==0.6.2
et-xmlfile==1.1.0
flake8==6.1.0
flake8-black==0.3.6
flake8-broken-line==1.0.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
frozenlist==1.4.1
idna==3.7
isort==5.13.2
Jinja2==3.1.4
kombu==5.3.5
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.0.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
paramiko==3.3.1
pathspec==0.12.1
pep8-naming==0.13.3
pexpect==4.8.0
platformdirs==4.2.2
ply==3.11
prompt-toolkit==3.0.43
propcache==0.2.0
ptyprocess==0.7.0
pyasn1==0.6.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.9.0
PyNaCl==1.5.0
pysmi==1.4.4
pysnmp==6.2.4
pysnmpcrypto==0.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rply==0.7.8
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.4.4
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
XlsxWriter==3.2.0
yarl==1.16.0
